+++ packages/zod-schemas/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./arrays";
 export * from "./principal";
+export * from "./result";
 export * from "./schema-id";
 export * from "./url";